refactor(core): replace RecordingSession with per-app SessionRegistry
Sessions are now one per app bundle ID, with stable monotonic UInt64
ids assigned at open and concurrent overlapping sessions supported as
views over one shared ring, per decision
docs/decisions/20260704-session-per-app.md. Also introduces:

- BufferPolicy: user-configurable ring size presets (5/10/30/60/120
  min) with resize-loss accounting (docs/decisions/20260704-configurable-ring-buffer.md).
- WaveformSampler: incremental peak sampling for session waveform
  display, finalized on session close.
- ByteRingBuffer seeded init + slice(lastBytes:), the primitives
  SessionRegistry.resize needs to preserve-newest-that-fits across a
  capacity change.

DumpSessionSelector, SessionSummary, and ListenerService are updated
throughout for UInt64 ids and newest-first ordering. ListenerService
keeps a transitional legacy-bundle-ID shim for the pre-CaptureOrchestrator
single-session callers; ListenerService.dump(sessionId:) also gains a
sessionNotFound code needed by the next commit's NDJSON error format.

+++ Sources/ManbokCore/Domain/SessionSummary.swift
@@ -1,8 +1,8 @@
 import Foundation
 
-/// One captured session slice in the ring (1-based id assigned at list time).
+/// One captured session slice in the ring (stable id assigned at session open).
 public struct SessionSummary: Sendable, Equatable {
-    public let id: Int
+    public let id: UInt64
     public let audioBytes: Int
     public let durationSeconds: Double
     /// Seconds before `now` that capture for this session began.
@@ -14,7 +14,7 @@ public struct SessionSummary: Sendable, Equatable {
     public let appName: String?
 
     public init(
-        id: Int,
+        id: UInt64,
         audioBytes: Int,
         durationSeconds: Double,
         startedSecondsAgo: TimeInterval,
